requirements.txt
# Python version (this is a comment - actual version control is done through runtime)
# Requires Python 3.11

# Core packages
streamlit==1.29.0
openai>=1.14.0
langchain==0.0.352
llama-index==0.12.9

# Document processing
python-docx==1.0.1
pypdf>=5.1.0,<6.0.0

# Development tools
python-dotenv==1.0.0
pytest==7.4.3

# Data processing
pandas==2.1.4
numpy==1.26.2
